[POAE7-2815] Add namespace and resolve redefinit.. (#333)
