+++ src/cider/common/hashtable/HashSet.h
@@ -30,7 +30,6 @@ namespace cider::hashtable {
 namespace ErrorCodes {
 extern const int LOGICAL_ERROR;
 }
-}  // namespace cider::hashtable
 
 /** NOTE HashSet could only be used for memmoveable (position independent) types.
  * Example: std::string is not position independent in libstdc++ with C++11 ABI or in
@@ -190,3 +189,5 @@ using HashSetWithSavedHash =
 //     HashTableAllocatorWithStackMemory<
 //         (1ULL << initial_size_degree)
 //         * sizeof(HashSetCellWithSavedHash<Key, Hash>)>>;
+
+}  // namespace cider::hashtable
