+++ src/cider/common/interpreters/AggregationHashTable.h
@@ -84,7 +84,6 @@ using AggregatedHashTableWithNullableUInt16Key =
     AggregatedHashTableWithNullKey<AggregatedHashTableWithUInt16Key>;
 
 class AggregationHashTable;
-HashTableAllocator allocator;
 
 struct AggregationMethod : private boost::noncopyable {
   enum class Type {
@@ -135,13 +134,7 @@ class AggregationHashTable final {
   // init_len: initial value length
   // `init_addr` and `init_len` describe the init value of a value in HashTable.
   // The memory layout can of any kind and should be designed by users.
-  explicit AggregationHashTable(std::vector<SQLTypes> key_types,
-                                int8_t* addr,
-                                uint32_t len)
-      : key_types_(key_types), init_val_(addr), init_len_(len) {
-    // TODO: use agg_method to construct the specific HashTable instead of all
-    agg_method_ = chooseAggregationMethod();
-  }
+  AggregationHashTable(std::vector<SQLTypes> key_types, int8_t* addr, uint32_t len);
 
   // raw_key: Layout of keys should be aligned to 16 like below:
   // |<-- key1_isNUll -->|<-- pad_1 -->|<-- key1_values -->|<-- key2_isNull -->| .....
@@ -150,89 +143,16 @@ class AggregationHashTable final {
   // |<-- v1_int8 -->|<-- pad -->| or |<-- v1_int32 -->| or |<-- v1_bool -->|<-- pad -->|
   // |<- 8bit ->|<- 8bit ->| or |<--- 32bit  --->| or |<- 8bit ->|<- 8bit ->|
   // return: start position of value
-  AggregateDataPtr get(int8_t* raw_key) {
-    // Transfer all keys to one AggKey
-    AggKey key = transferToAggKey(raw_key);
-    // key_set_.emplace(key);
+  AggregateDataPtr get(int8_t* raw_key);
 
-    for (int i = 0; i < key_types_.size(); i++) {
-      if (SQLTypes::kTINYINT == key_types_[i]) {
-        uint8_t key_v = (reinterpret_cast<uint8_t*>(key.getAddr()))[0];
-        if (agg_ht_uint8_[key_v] == nullptr) {
-          // Allocate memory of values here since value type like non-fixed length address
-          // cannot be new in hash table. This case should be manually handled and it's
-          // better to use an Arena for better memory efficiency.
-          agg_ht_uint8_[key_v] = allocator.allocate(init_len_);
-          std::memcpy(agg_ht_uint8_[key_v], init_val_, init_len_);
-        }
-        return agg_ht_uint8_[key_v];
-      } else if (SQLTypes::kSMALLINT == key_types_[i]) {
-        uint16_t key_v = (reinterpret_cast<uint16_t*>(key.getAddr()))[0];
-        if (agg_ht_uint16_[key_v] == nullptr) {
-          // Allocate memory of values here since value type like non-fixed length address
-          // cannot be new in hash table. This case should be manually handled and it's
-          // better to use an Arena for better memory efficiency.
-          agg_ht_uint16_[key_v] = allocator.allocate(init_len_);
-          std::memcpy(agg_ht_uint16_[key_v], init_val_, init_len_);
-        }
-        return agg_ht_uint16_[key_v];
-      } else if (SQLTypes::kINT == key_types_[i]) {
-        uint32_t key_v = (reinterpret_cast<uint32_t*>(key.getAddr()))[0];
-        if (agg_ht_uint32_[key_v] == nullptr) {
-          agg_ht_uint32_[key_v] = allocator.allocate(init_len_);
-          std::memcpy(agg_ht_uint32_[key_v], init_val_, init_len_);
-        }
-        return agg_ht_uint32_[key_v];
-      } else if (SQLTypes::kBIGINT == key_types_[i]) {
-        uint64_t key_v = (reinterpret_cast<uint64_t*>(key.getAddr()))[0];
-        if (agg_ht_uint64_[key_v] == nullptr) {
-          agg_ht_uint64_[key_v] = allocator.allocate(init_len_);
-          std::memcpy(agg_ht_uint64_[key_v], init_val_, init_len_);
-        }
-        return agg_ht_uint64_[key_v];
-      }
-    }
-    CIDER_THROW(CiderRuntimeException, "Unsupported key type");
-  }
-
-  AggregateDataPtr get(std::vector<AggKey> agg_keys) {
-    CIDER_THROW(CiderRuntimeException, "Unsupported key type");
-  }
+  AggregateDataPtr get(std::vector<AggKey> agg_keys);
 
   // key_addr: Same as `key_addr` in get
   // return: AggKey stored in HashTable
   // This function is to transfer keys formatted in codegen to AggKey in HashTable.
   // It will try to merge keys to a primitive type in multiple key cases.
   // If failed, serialize all keys to one key in Type::SERIALIZED.
-  AggKey transferToAggKey(int8_t* key_addr) {
-    // Single key
-    if (1 == key_types_.size()) {
-      if (SQLTypes::kTINYINT == key_types_[0]) {
-        bool is_null = (reinterpret_cast<bool*>(key_addr))[0];
-        AggKey key(is_null, key_addr + 2, 1);
-        return key;
-      }
-      if (SQLTypes::kSMALLINT == key_types_[0]) {
-        bool is_null = (reinterpret_cast<bool*>(key_addr))[0];
-        AggKey key(is_null, key_addr + 2, 2);
-        return key;
-      }
-      if (SQLTypes::kINT == key_types_[0]) {
-        bool is_null = (reinterpret_cast<bool*>(key_addr))[0];
-        AggKey key(is_null, key_addr + 2, 4);
-        return key;
-      }
-      if (SQLTypes::kBIGINT == key_types_[0]) {
-        bool is_null = (reinterpret_cast<bool*>(key_addr))[0];
-        AggKey key(is_null, key_addr + 2, 8);
-        return key;
-      }
-      // TODO: Support more types.
-    }
-    CIDER_THROW(CiderRuntimeException, "Unsupported Aggregation key");
-    // TODO: Multiple keys, find out if keys can be arranged to primitive types like
-    // int32/int64... If not, serialize the key and set the key type to Type::serialized.
-  }
+  AggKey transferToAggKey(int8_t* key_addr);
 
   // Dump all value of the HashTable.
   // TODO: Here need to be discussed, what to return?
@@ -264,22 +184,6 @@ class AggregationHashTable final {
   AggregatedHashTableWithUInt64Key agg_ht_uint64_;
 
   // Select the aggregation method based on the number and types of keys.
-  AggregationMethod::Type chooseAggregationMethod() {
-    // Single key
-    if (1 == key_types_.size()) {
-      if (SQLTypes::kTINYINT == key_types_[0]) {
-        return AggregationMethod::Type::INT8;
-      } else if (SQLTypes::kSMALLINT == key_types_[0]) {
-        return AggregationMethod::Type::INT16;
-      } else if (SQLTypes::kINT == key_types_[0]) {
-        return AggregationMethod::Type::INT32;
-      } else if (SQLTypes::kBIGINT == key_types_[0]) {
-        return AggregationMethod::Type::INT64;
-      }
-      // TODO: Support more types.
-    }
-    // TODO: Support multiple keys.
-    return AggregationMethod::Type::EMPTY;
-  }
+  AggregationMethod::Type chooseAggregationMethod();
 };
 }  // namespace cider::hashtable
